# FleetSync-Vehicle-Services/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vehicle Service API",
        "description": "API documentation for Vehicle Service",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get all vehicles",
                "operationId": "727634867f59796290a2fff276979198",
                "responses": {
                    "200": {
                        "description": "List of vehicles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "Car"
                                                    },
                                                    "plate_number": {
                                                        "type": "string",
                                                        "example": "B 1234 ABC"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Available"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Create a new vehicle",
                "operationId": "a88e0bca4193c91e713d0441659d199c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "plate_number",
                                    "status"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Car"
                                    },
                                    "plate_number": {
                                        "type": "string",
                                        "example": "B 1234 ABC"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Available",
                                            "InUse",
                                            "Maintenance"
                                        ],
                                        "example": "Available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vehicle created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Car"
                                                },
                                                "plate_number": {
                                                    "type": "string",
                                                    "example": "B 1234 ABC"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Available"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get a specific vehicle",
                "operationId": "0185374b4b707b85ca8a97a6f5f366e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Car"
                                                },
                                                "plate_number": {
                                                    "type": "string",
                                                    "example": "B 1234 ABC"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Available"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Update a vehicle",
                "operationId": "c4d991b5c398a7ccd4e73f9a81c355d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Car"
                                    },
                                    "plate_number": {
                                        "type": "string",
                                        "example": "B 1234 ABC"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Available",
                                            "InUse",
                                            "Maintenance"
                                        ],
                                        "example": "Available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vehicle updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Car"
                                                },
                                                "plate_number": {
                                                    "type": "string",
                                                    "example": "B 1234 ABC"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Available"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Delete a vehicle",
                "operationId": "3442f99c22f47affb335443ef012e712",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vehicle deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Vehicles",
            "description": "Vehicles"
        }
    ]
}